spacing on homepage categories
